Reorganising some general parameters and function
