+++ views/general.py
@@ -162,3 +162,12 @@ def cache_on_browser(minutes=5):
         return wrapped_f
 
     return fwrap
+
+
+def _get_pagination_parameters():
+    try:
+        offset = int(request.args.get("offset", 0))
+        limit = int(request.args.get("limit", 10))
+    except ValueError as e:
+        raise PhenopolisException(str(e), 500)
+    return limit, offset
